requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

alembic==1.5.2
azure-functions
azure-identity
azure-keyvault
azure-keyvault-keys
azure-storage-blob
bcrypt==3.2.0
firebase-admin==4.5.1
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-magic==0.4.22
sendgrid==6.6.0
SQLAlchemy==1.3.22
validator==0.7.0